Refactor default data handler
