+++ src/org/traccar/handler/DefaultDataHandler.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2015 Anton Tananaev (anton@traccar.org)
+ * Copyright 2015 - 2019 Anton Tananaev (anton@traccar.org)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,21 +13,31 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.traccar;
+package org.traccar.handler;
 
+import io.netty.channel.ChannelHandler;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.traccar.BaseDataHandler;
+import org.traccar.database.DataManager;
 import org.traccar.model.Position;
 
+@ChannelHandler.Sharable
 public class DefaultDataHandler extends BaseDataHandler {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(DefaultDataHandler.class);
+
+    private final DataManager dataManager;
+
+    public DefaultDataHandler(DataManager dataManager) {
+        this.dataManager = dataManager;
+    }
 
     @Override
     protected Position handlePosition(Position position) {
 
         try {
-            Context.getDataManager().addObject(position);
+            dataManager.addObject(position);
         } catch (Exception error) {
             LOGGER.warn("Failed to store position", error);
         }
